refactor(utils): change name
