+++ src/test/utils/randomKoreanWord.spec.js
@@ -1,8 +1,8 @@
-import randomWord from "../../utils/randomWord";
+import randomKoreanWord from "../../utils/randomKoreanWord";
 
 describe("random word", () => {
   it("should return true when random word is generated", () => {
-    const word = randomWord();
+    const word = randomKoreanWord();
     expect(word.length > 0).toBeTruthy();
   });
 });
