+++ src/utils/index.js
@@ -1,12 +1,13 @@
-export { default as dateToYYYYMMDD } from "./dateToYYYYMMDD.js";
-export { default as isFunction } from "./isFunction.js";
-export { default as isPlainObject } from "./isPlainObject.js";
-export { default as isValidDate } from "./isValidDate.js";
-export { default as oneOf } from "./oneOf.js";
-export { default as randomDate } from "./randomDate.js";
-export { default as randomImage } from "./randomImage.js";
-export { default as randomKoreanAddress } from "./randomKoreanAddress.js";
-export { default as randomKoreanName } from "./randomKoreanName.js";
-export { default as randomMoney } from "./randomMoney.js";
-export { default as randomNumber } from "./randomNumber.js";
-export { default as randomWord } from "./randomWord.js";
+export { default as dateToYYYYMMDD } from "./dateToYYYYMMDD";
+export { default as isFunction } from "./isFunction";
+export { default as isPlainObject } from "./isPlainObject";
+export { default as isValidDate } from "./isValidDate";
+export { default as oneOf } from "./oneOf";
+export { default as randomDate } from "./randomDate";
+export { default as randomImage } from "./randomImage";
+export { default as randomKoreanAddress } from "./randomKoreanAddress";
+export { default as randomKoreanName } from "./randomKoreanName";
+export { default as randomMoney } from "./randomMoney";
+export { default as randomNumber } from "./randomNumber";
+export { default as randomKoreanWord } from "./randomKoreanWord";
+export { default as randomSentenceWord } from "./randomKoreanSentence";
