+++ src/utils/randomKoreanSentence.js
@@ -15,7 +15,7 @@ const sizeTable = {
  * @param { 'short' | 'medium' | 'long' } Size of generated sentence.
  * @returns { string } Generated sentence of a given size.
  */
-const randomSentence = (size = "short") => {
+const randomKoreanSentence = (size = "short") => {
   const sentenceCount = sizeTable[size];
   if (!sentenceCount) {
     throw new Error("size should be one of short, medium, long");
@@ -23,4 +23,4 @@ const randomSentence = (size = "short") => {
   return sample(sentences, sentenceCount).join(". ") + "."; // 비용을 아끼기 위해서 불변성을 유지하지 않는다.
 };
 
-export default randomSentence;
+export default randomKoreanSentence;
